qPCR data sets by Karlen et al. (2007)

One single tabular tidy data set in long format, encompassing 32 dilution series, for seven PCR targets and four biological samples. The targeted amplicons are within the murine genes: Cav1, Ccn2, Eln, Fn1, Rpl27, Hspg2, and Serpine1, respectively. Dilution series: scheme 1 (Cav1, Eln, Hspg2, Serpine1): 1-fold, 10-fold, 50-fold, and 100-fold; scheme 2 (Ccn2, Rpl27, Fn1): 1-fold, 10-fold, 50-fold, 100-fold and 1000-fold. For each concentration there are five replicates, except for the 1000-fold concentration, where only two replicates were performed. Each amplification curve is 40 cycles long. Please read the sections Experimental set of qPCR data and Quantitative PCR assays of Karlen et al. (2007) for more details.

Format

A tibble providing amplification curve data in long format.

plate

Plate identifier. This corresponds, loosely, to the name of the targets. In the original publication the amplicons are indicated by gene symbol synonyms which we do use here for naming each plate. This differs from the names used in the column target where actual murine gene symbols are used.

well

Well identifier, i.e. the position within the 96-well plate.

dye

Type of fluorescence dye, in this data set it is always SYBR Green I master mix (Roche) ("SYBR").

target

Target identifier: murine gene symbol where the amplicon maps to.

sample_type

Sample type. Most reactions in this data set are standard curves, i.e. "std", but a few no template controls ("ntc") are also included.

replicate

Replicate identifier: 1 thru 5.

dilution

Dilution factor. Higher number means greater dilution, e.g. 10 means a 1:10 (ten-fold) dilution.

cycle

PCR cycle.

fluor

Raw fluorescence values.

Source

doi:10.1186/1471-2105-8-131

Examples

karlen
#> # A tibble: 28,680 × 11
#>    plate well  target dye   sample sample_type replicate copies dilution cycle
#>    <fct> <fct> <fct>  <fct> <fct>  <fct>       <fct>      <int>    <int> <dbl>
#>  1 CAV   A1    Cav1   SYBR  S1     std         1             NA        1     1
#>  2 CAV   A1    Cav1   SYBR  S1     std         1             NA        1     2
#>  3 CAV   A1    Cav1   SYBR  S1     std         1             NA        1     3
#>  4 CAV   A1    Cav1   SYBR  S1     std         1             NA        1     4
#>  5 CAV   A1    Cav1   SYBR  S1     std         1             NA        1     5
#>  6 CAV   A1    Cav1   SYBR  S1     std         1             NA        1     6
#>  7 CAV   A1    Cav1   SYBR  S1     std         1             NA        1     7
#>  8 CAV   A1    Cav1   SYBR  S1     std         1             NA        1     8
#>  9 CAV   A1    Cav1   SYBR  S1     std         1             NA        1     9
#> 10 CAV   A1    Cav1   SYBR  S1     std         1             NA        1    10
#> # ℹ 28,670 more rows
#> # ℹ 1 more variable: fluor <dbl>